+++ sklearn/utils/graph.py
@@ -112,7 +112,7 @@ def graph_laplacian(csgraph, normed=False, return_diag=False):
     if csgraph.ndim != 2 or csgraph.shape[0] != csgraph.shape[1]:
         raise ValueError('csgraph must be a square matrix or array')
 
-    if normed and (np.issubdtype(csgraph.dtype, np.int)
+    if normed and (np.issubdtype(csgraph.dtype, int)
                    or np.issubdtype(csgraph.dtype, np.uint)):
         csgraph = check_array(csgraph, dtype=np.float64, accept_sparse=True)
